Move utils to its own workspace
This is to be able to reuse them from integrations tests

automated_actions_utils is not defined as a workspace member in the root
pyproject.toml so that we can express dependencies from it. If defined
as a workspace member in the root pyproject, dependencies via workspace
don't work in the automated_actions pyproject and nested workspaces are
not allowed in uv.

Signed-off-by: Rafa Porres Molina <rporresm@redhat.com>